folderStructureAdded
